# Close #69 Optimize the `incl_prefix_sum` function when particles do not fill the grid