--- a/fbpic/particles/particles.py
+++ b/fbpic/particles/particles.py
@@ -967,7 +967,8 @@ class Particles(object) :
         sort_particles_per_cell(self.cell_idx, self.sorted_idx)
         # Reset the old prefix sum
         fld.prefix_sum_shift = 0
-        prefill_prefix_sum[dim_grid_2d_flat, dim_block_2d_flat](self.prefix_sum)
+        prefill_prefix_sum[dim_grid_2d_flat, dim_block_2d_flat](
+            self.cell_idx, self.prefix_sum)
         # Perform the inclusive parallel prefix sum
         incl_prefix_sum[dim_grid_1d, dim_block_1d](
             self.cell_idx, self.prefix_sum)