This pull request makes the inclusive prefix sum more efficient, by avoiding that the thread handling the last particle (i.e. the rightmost particle) fills all the cells that are to its right. (see the code for a clearer picture).

This is important when a species (e.g. an electron bunch) fills only a small fraction of the plasma, since then the thread handling the rightmost particle has to fill a significant fraction of the box. But it is also important in regular simulations with guard cells, since then the thread for the rightmost particles needs to fill a significant fraction of the guard cells.

I did a few tests with a 2000x400 box with moving window (and thus guard cells) using nvprof:

- With a plasma filling the valid cells (but not the guard cells):
   
   - Current dev branch: `incl_prefix_sum` takes 3.46% of the kernel time (1.71583s)
   - This PR: `incl_prefix_sum` takes 0.40%  (192.88ms)

- With a bunch filling only a small fraction of the grid:

   - Current dev branch: `incl_prefix_sum` takes 13.64%  (4.64948s)
   - This PR: `incl_prefix_sum` takes 0.40%  (117.90ms)

In all cases, both the old `reset_prefix_sum` and the new `prefill_prefix_sum` take a negligible amount of time.