@RemiLehe 2017-08-05
Thanks! 
By the way, I just ran the automated tests on GPU for this branch: they all pass.